Adequação de fechamento_turma para base de consulta
