+++ src/SME.SGP.Aplicacao/Queries/FechamentoNota/ObterAlunosComNotaLancadaPorPeriodoEscolarUE/ObterAlunosComNotaLancadaPorPeriodoEscolarUEQuery.cs
@@ -1,8 +1,6 @@
 ﻿using FluentValidation;
 using MediatR;
-using SME.SGP.Dominio;
 using SME.SGP.Infra;
-using System;
 using System.Collections.Generic;
 
 namespace SME.SGP.Aplicacao
