+++ src/SME.SGP.Dominio.Servicos/ServicoPendenciaFechamento.cs
@@ -1,5 +1,4 @@
 ﻿using MediatR;
-using SME.Background.Core;
 using SME.SGP.Aplicacao;
 using SME.SGP.Dominio.Interfaces;
 using SME.SGP.Infra;
@@ -19,7 +18,7 @@ namespace SME.SGP.Dominio.Servicos
         private readonly IRepositorioAula repositorioAula;
         private readonly IRepositorioPendencia repositorioPendencia;
         private readonly IRepositorioPendenciaFechamento repositorioPendenciaFechamento;
-        private readonly IRepositorioFechamentoNota repositorioFechamentoNota;
+        private readonly IRepositorioFechamentoNotaConsulta repositorioFechamentoNota;
         private readonly IServicoUsuario servicoUsuario;
         private readonly IServicoAbrangencia servicoAbrangencia;
         private readonly IRepositorioComponenteCurricularConsulta repositorioComponenteCurricular;
@@ -38,7 +37,7 @@ namespace SME.SGP.Dominio.Servicos
                                           IRepositorioPendenciaFechamento repositorioPendenciaFechamento,
                                           IRepositorioAula repositorioAula,
                                           IRepositorioComponenteCurricularConsulta repositorioComponenteCurricular,
-                                          IRepositorioFechamentoNota repositorioFechamentoNota,
+                                          IRepositorioFechamentoNotaConsulta repositorioFechamentoNota,
                                           IServicoUsuario servicoUsuario,
                                           IServicoAbrangencia servicoAbrangencia,
                                           IMediator mediator)
